requirements.txt
pandas
numpy
matplotlib
seaborn
python-dotenv
yfinance
ccxt
schedule
pytz
